+++ java-sdk/src/main/java/com/baidu/dueros/data/request/RequestBody.java
@@ -17,11 +17,10 @@
 package com.baidu.dueros.data.request;
 
 import com.baidu.dueros.data.request.events.ButtonClickedEvent;
+import com.baidu.dueros.data.request.events.ConnectionsResponseEvent;
 import com.baidu.dueros.data.request.events.ElementSelectedEvent;
-import com.baidu.dueros.data.request.events.LinkAccountSucceededEvent;
 import com.baidu.dueros.data.request.events.LinkClickedEvent;
 import com.baidu.dueros.data.request.events.RadioButtonClickedEvent;
-import com.baidu.dueros.data.request.pay.event.ChargeEvent;
 import com.baidu.dueros.data.request.permission.event.PermissionGrantFailedEvent;
 import com.baidu.dueros.data.request.permission.event.PermissionGrantedEvent;
 import com.baidu.dueros.data.request.permission.event.PermissionRejectedEvent;
@@ -72,11 +71,10 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
         @Type(value = PlaybackStutterStartedEvent.class), @Type(value = PlaybackStutterFinishedEvent.class),
         @Type(value = PlaybackPausedEvent.class), @Type(value = PlaybackResumedEvent.class),
         @Type(value = PlaybackQueueClearedEvent.class), @Type(value = ElementSelectedEvent.class),
-        @Type(value = LinkClickedEvent.class), @Type(value = ChargeEvent.class),
+        @Type(value = LinkClickedEvent.class), @Type(value = ConnectionsResponseEvent.class),
         @Type(value = PermissionGrantFailedEvent.class), @Type(value = PermissionRejectedEvent.class),
         @Type(value = PermissionGrantedEvent.class), @Type(value = PermissionRequiredEvent.class),
-        @Type(value = LinkAccountSucceededEvent.class), @Type(value = ButtonClickedEvent.class),
-        @Type(value = RadioButtonClickedEvent.class)})
+        @Type(value = ButtonClickedEvent.class), @Type(value = RadioButtonClickedEvent.class) })
 @JsonIgnoreProperties(ignoreUnknown = true)
 public abstract class RequestBody {